Working on listing topics in UI
More work on directly attaching to topics

Annotates autocreated topic config maps so that intermediate topics
do not appear in the editor menu.

Working on listing topics in UI

More work on directly attaching to topics

Working on listing topics in UI

Annotates autocreated topic config maps so that intermediate topics
do not appear in the editor menu.